python-hjy/requirements.txt
Flask==1.1.1
Flask-Sockets==0.2.1
gevent-websocket==0.10.1
mysql-connector-python==8.0.19
pyahocorasick==1.4.0